+++ src/App.tsx
@@ -7,8 +7,11 @@ const App = () => (
       <GoogleMapProvider>
         <MapBox
           apiKey={process.env.REACT_APP_GOOGLE_MAP_API_KEY || ''}
-          center={{lat: 39, lng: 116}}
-          zoom={14}
+          opts={{
+            center: {lat: 39, lng: 116},
+            noClear: true,
+            zoom: 14,
+          }}
           useDrawing={true}
           useGeometry={true}
           usePlaces={true}
@@ -17,10 +20,18 @@ const App = () => (
             console.log('The center of the map has changed.')
           }}
         />
-        <Marker draggable label="hello" position={{lat: 39, lng: 116}} />
+        <Marker
+          opts={{
+            draggable: true,
+            label: 'hello',
+            position: {lat: 39, lng: 116},
+          }}
+        />
         <InfoWindow
-          content="This is an info window"
-          position={{lat: 39.01, lng: 115.99}}
+          opts={{
+            content: 'This is an info window',
+            position: {lat: 39.01, lng: 115.99},
+          }}
           visible
         />
         <Polygon
@@ -29,6 +40,7 @@ const App = () => (
             {lat: 38.98, lng: 116.03},
             {lat: 38.99, lng: 116.03},
           ]}
+          visible
         />
       </GoogleMapProvider>
     </header>