+++ src/lib/__test__helpers__/index.ts
@@ -1,1 +1,2 @@
 export {default as defineGlobalVariable} from './defineGlobalVariable'
+export {default as FakeComponent} from './FakeComponent'