+++ src/lib/components/__tests__/Polygon.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import 'jest-dom/extend-expect'
 import loadjs from 'loadjs'
 import 'react-testing-library/cleanup-after-each'
-import {render, wait, cleanup} from 'react-testing-library'
+import {render, wait, cleanup, flushEffects} from 'react-testing-library'
 import {Polygon, MapBox} from '../..'
 import {GoogleMapProvider} from '../../contexts/GoogleMapContext'
 import {defineGlobalVariable} from '../../__test__helpers__'
@@ -20,7 +20,7 @@ describe('Polygon', () => {
   })
 
   it('renders inside a MapBox', async () => {
-    const {container} = render(
+    const {container, rerender} = render(
       <GoogleMapProvider>
         <MapBox apiKey="A_FAKE_API_KEY" />
         <Polygon
@@ -28,11 +28,17 @@ describe('Polygon', () => {
         />
       </GoogleMapProvider>,
     )
-    expect(container.innerHTML).toMatch('Loading...')
     await wait(() => {
       expect(container.innerHTML).not.toMatch('Loading...')
     })
-    expect(loadjs.reset).not.toHaveBeenCalled()
-    expect(container.innerHTML).toMatch('This is a map')
+    rerender(
+      <GoogleMapProvider>
+        <MapBox apiKey="A_FAKE_API_KEY" />
+        <Polygon
+          paths={[{lat: 31, lng: 18}, {lat: 36, lng: 19}, {lat: 39, lng: 20}]}
+        />
+      </GoogleMapProvider>,
+    )
+    flushEffects()
   })
 })