+++ src/lib/contexts/__tests__/GoogleMapContext.test.tsx
@@ -1,31 +1,22 @@
-import React, {useContext, useEffect} from 'react'
-import {GoogleMapContext, GoogleMapProvider} from '../GoogleMapContext'
+import React from 'react'
+import {GoogleMapProvider} from '../GoogleMapContext'
 import 'react-testing-library/cleanup-after-each'
 import {render, cleanup, flushEffects} from 'react-testing-library'
-import {defineGlobalVariable} from '../../__test__helpers__'
-import {GoogleMapAction} from '../../common/types'
+import {defineGlobalVariable, FakeComponent} from '../../__test__helpers__'
 
-const FakeComponent = ({action}: {action: GoogleMapAction}) => {
-  const {state, dispatch} = useContext(GoogleMapContext)
-  useEffect(() => {
-    dispatch(action)
-  }, [state])
-  return <div />
-}
+beforeEach(() => {
+  defineGlobalVariable()
+  jest.spyOn(console, 'error')
+})
 
-describe('The dispatcher', () => {
-  beforeEach(() => {
-    defineGlobalVariable()
-    jest.spyOn(console, 'error')
-  })
+afterEach(() => {
+  cleanup()
+  Object.defineProperty(global, 'google', {value: undefined})
+  jest.restoreAllMocks()
+})
 
-  afterEach(() => {
-    cleanup()
-    Object.defineProperty(global, 'google', {value: undefined})
-    jest.restoreAllMocks()
-  })
-
-  it('throws an error when trying to init map without a map instance', () => {
+describe('The dispatcher throws an error when trying to', () => {
+  it('init map without a map instance', () => {
     expect(() => {
       render(
         <GoogleMapProvider>
@@ -36,7 +27,7 @@ describe('The dispatcher', () => {
     }).toThrowError(new Error('You should specify a map instance'))
   })
 
-  it('throws an error when trying to add more than one map to one context', () => {
+  it('add more than one map to one context', () => {
     expect(() => {
       const map = new google.maps.Map(document.createElement('div'), {zoom: 14})
       render(
@@ -51,7 +42,7 @@ describe('The dispatcher', () => {
     )
   })
 
-  it('throws an error when trying to add a marker without a marker instance', () => {
+  it('add a marker without a marker instance', () => {
     expect(() => {
       render(
         <GoogleMapProvider>
@@ -62,7 +53,7 @@ describe('The dispatcher', () => {
     }).toThrowError(new Error('You should specify a marker instance'))
   })
 
-  it('throws an error when trying to add the same marker', () => {
+  it('add the same marker more than once', () => {
     expect(() => {
       const marker = new google.maps.Marker({position: {lat: 0, lng: 0}})
       render(
@@ -75,7 +66,7 @@ describe('The dispatcher', () => {
     }).toThrowError(new Error('The marker has already been added'))
   })
 
-  it('throws an error when trying to remove a marker without a marker instance', () => {
+  it('remove a marker without a marker instance', () => {
     expect(() => {
       render(
         <GoogleMapProvider>
@@ -98,9 +89,58 @@ describe('The dispatcher', () => {
     }).toThrowError(new Error('The marker cannot be found'))
   })
 
-  it('can reset', () => {
-    const marker = new google.maps.Marker({position: {lat: 0, lng: 0}})
+  it('add a polygon without a polygon instance', () => {
     expect(() => {
+      render(
+        <GoogleMapProvider>
+          <FakeComponent action={{type: 'add_polygon'}} />
+        </GoogleMapProvider>,
+      )
+      flushEffects()
+    }).toThrowError(new Error('You should specify a polygon instance'))
+  })
+
+  it('add the same polygon more than once', () => {
+    expect(() => {
+      const polygon = new google.maps.Polygon({})
+      render(
+        <GoogleMapProvider>
+          <FakeComponent action={{type: 'add_polygon', polygon: polygon}} />
+          <FakeComponent action={{type: 'add_polygon', polygon: polygon}} />
+        </GoogleMapProvider>,
+      )
+      flushEffects()
+    }).toThrowError(new Error('The polygon has already been added'))
+  })
+
+  it('remove a polygon without a polygon instance', () => {
+    expect(() => {
+      render(
+        <GoogleMapProvider>
+          <FakeComponent action={{type: 'remove_polygon'}} />
+        </GoogleMapProvider>,
+      )
+      flushEffects()
+    }).toThrowError(new Error('You should specify a polygon instance'))
+  })
+
+  it('throws an error when trying to remove a non-existing polygon', () => {
+    expect(() => {
+      const polygon = new google.maps.Polygon({})
+      render(
+        <GoogleMapProvider>
+          <FakeComponent action={{type: 'remove_polygon', polygon: polygon}} />
+        </GoogleMapProvider>,
+      )
+      flushEffects()
+    }).toThrowError(new Error('The polygon cannot be found'))
+  })
+})
+
+describe('The dispatcher will', () => {
+  it('reset the map', () => {
+    expect(() => {
+      const marker = new google.maps.Marker({position: {lat: 0, lng: 0}})
       render(
         <GoogleMapProvider>
           <FakeComponent action={{type: 'add_marker', marker: marker}} />
@@ -111,7 +151,7 @@ describe('The dispatcher', () => {
     }).not.toThrow()
   })
 
-  it('can remove a marker', () => {
+  it('add and remove a marker', () => {
     expect(() => {
       const marker = new google.maps.Marker({position: {lat: 0, lng: 0}})
       render(
@@ -124,7 +164,20 @@ describe('The dispatcher', () => {
     }).not.toThrow()
   })
 
-  it('ignores undefined action types', () => {
+  it('add and remove a polygon', () => {
+    expect(() => {
+      const polygon = new google.maps.Polygon({})
+      render(
+        <GoogleMapProvider>
+          <FakeComponent action={{type: 'add_polygon', polygon: polygon}} />
+          <FakeComponent action={{type: 'remove_polygon', polygon: polygon}} />
+        </GoogleMapProvider>,
+      )
+      flushEffects()
+    }).not.toThrow()
+  })
+
+  it('ignore undefined action types', () => {
     expect(() => {
       render(
         <GoogleMapProvider>