+++ src/stories/components/MarkerDraggable.tsx
@@ -6,7 +6,7 @@ const scope = {GoogleMapProvider, MapBox, Marker}
 
 const code = `<GoogleMapProvider>
   <MapBox apiKey="" LoadedComponent={() => <h1>The marker can be draggable</h1>} />
-  <Marker position={{lat: 39, lng: 116}} draggable />
+  <Marker opts={{draggable: true, position: {lat: 39, lng: 116}}} />
 </GoogleMapProvider>
 `