+++ src/stories/components/WithInfoWindow.tsx
@@ -7,8 +7,10 @@ const scope = {GoogleMapProvider, InfoWindow, MapBox}
 const code = `<GoogleMapProvider>
   <MapBox apiKey="" LoadedComponent={() => <h1>Display an info window</h1>} />
   <InfoWindow 
-    content="This is an info window"
-    position={{lat: 39, lng: 116}}
+    opts={{
+      content: 'This is an info window',
+      position: {lat: 39, lng: 116}
+    }}
     visible
   />
 </GoogleMapProvider>