+++ src/stories/components/WithMarker.tsx
@@ -6,7 +6,7 @@ const scope = {GoogleMapProvider, MapBox, Marker}
 
 const code = `<GoogleMapProvider>
   <MapBox apiKey="" LoadedComponent={() => <h1>Display a marker</h1>} />
-  <Marker position={{lat: 39, lng: 116}} />
+  <Marker opts={{position: {lat: 39, lng: 116}}} />
 </GoogleMapProvider>
 `